refactor(data): query MongoDB through collections, not raw commands
Every question this application asked was a command document it assembled
itself: `aggregate` with its cursor options, `insert` with `ordered`,
`createIndexes` with both specifications, `drop`, and a `$count`
pipeline read back by hand. The library now carries all of that, so the
data layer keeps the part that is about coffee places and hands the rest
over.

`MongoSeam` is gone. `:data` depends on `embedded-mongodb-core` -- the
library's Android-free half -- and a `PlaceRepository` is built from a
`MongoCollection`, a `Seeder` from two. The command builders become
pipeline builders: `nearestPipeline`, `searchPipeline`,
`viewportPipeline` and `placeIndexes` are what this application knows and
nothing more. Seeding inserts with `insertMany(ordered = false)`, counts
with `countDocuments`, and writes its marker with `insertOne`.

The tests move down with them. `FakeMongo` is now a `CommandRunner`, so
what a test scripts is the seam the library itself is written against
rather than one this application invented, and the assertions are on the
commands the library builds around a pipeline.

`SeedProgress` counts are `Long`, because `countDocuments` reports one
and narrowing it would be a truncation nobody asked for.

`:data` still needs no emulator and no compiled engine; it does now need
the library checkout, because the core module is where the collections
live.

+++ app/src/main/kotlin/io/github/jeroenvervaeke/coffeefinder/ui/Timings.kt
@@ -85,7 +85,7 @@ class StartupTimer(
         phaseBegan = clock.markNow()
     }
 
-    private fun finish(places: Int) {
+    private fun finish(places: Long) {
         report()
         // Cleared so a seeder that somehow published twice cannot report the last phase twice.
         phase = null
